Moved Status to models, not models.status
