+++ src/main/kotlin/com/github/woojiahao/models/Status.kt
@@ -1,4 +1,4 @@
-package com.github.woojiahao.models.status
+package com.github.woojiahao.models
 
 sealed class Status<T>(val value: T) {
   class Fail<T>(default: T, vararg errorComponents: String): Status<T>(default) {
